Getting help and stop for table selection stage to prevent progression to the DML selection stage

+++ src/main/java/com/qa/ims/InventoryManager.java
@@ -25,60 +25,71 @@ public class InventoryManager {
 
 	public static final Logger LOGGER = Logger.getLogger(InventoryManager.class);
 
-	public void initiateSystem() {
+	public String initiateSystem() {
 
 		LOGGER.info("Please insert username: ");
 		String username = Utils.getInput();
 		LOGGER.info("Please insert password: ");
 		String password = Utils.getInput();
 
-		LOGGER.info("Hello, " + username + " how can I help you today?");
-		LOGGER.info("1): Review Customers      2): Review Products      3): Review Orders      4): Help      5): Stop");
+		LOGGER.info("Hello, " + username + ", how can I help you today?");
+		LOGGER.info(
+				"[1]: Review [customers]      [2]: Review [products]      [3]: Review [orders]      [4]: [help]      [5]: [stop]");
 		TableSelectCommand tableSelectCommand = TableSelectCommand.getTableSelectCommand();
 
 		LOGGER.info("You have selected '" + tableSelectCommand + "'. How would you like to proceed?");
-		LOGGER.info("1): Create      2): Read      3): Update      4): Delete      5): Help      6): Back");
+		LOGGER.info(
+				"[1]: [create]      [2]: [read]      [3]: [update]      [4]: [delete]      [5]: [help]      [6]: [back]");
 		DmlCommands dmlCommand = DmlCommands.getDmlCommands();
 
 		switch (tableSelectCommand) {
-		case CUSTOMER:
+		case CUSTOMERS:
 			CustomerController customerController = new CustomerController(
 					new CustomerServices(new CustomerDataAccessObject(username, password)));
 			doDMLCommand(customerController, dmlCommand);
 			break;
-		case PRODUCT:
+		case PRODUCTS:
 			ProductController productController = new ProductController(
 					new ProductServices(new ProductDataAccessObject(username, password)));
 			doDMLCommand(productController, dmlCommand);
 			break;
-		case ORDER:
+		case ORDERS:
+			break;
+		case HELP:
+			System.out.println("SDfvdsge");
+			TableSelectCommand.getTableSelectCommand();
 			break;
 		case STOP:
+			System.exit(0);
 			break;
 		default:
 			break;
 		}
+		return tableSelectCommand.toString();
 	}
 
 	public void doDMLCommand(CrudController<?> createReadUpdateDestroyController, DmlCommands dmlCommand) {
-		switch (dmlCommand) {
-		case CREATE:
-			createReadUpdateDestroyController.create();
-			break;
-		case READ:
-			createReadUpdateDestroyController.readAll();
-			break;
-		case UPDATE:
-			createReadUpdateDestroyController.update();
-			break;
-		case DELETE:
-			createReadUpdateDestroyController.delete();
-			break;
-		case RETURN:
-			break;
-		default:
-			break;
+		InventoryManager initiateDMLCommand = new InventoryManager();
+		if (initiateDMLCommand.initiateSystem() != "HELP") {
+			switch (dmlCommand) {
+			case CREATE:
+				createReadUpdateDestroyController.create();
+				break;
+			case READ:
+				createReadUpdateDestroyController.readAll();
+				break;
+			case UPDATE:
+				createReadUpdateDestroyController.update();
+				break;
+			case DELETE:
+				createReadUpdateDestroyController.delete();
+				break;
+			case RETURN:
+				break;
+			default:
+				break;
 
+			}
 		}
 	}
